requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Image processing
Pillow>=10.0.0

# Data handling
pandas>=2.0.0
tqdm>=4.65.0

# Web framework
Flask>=2.3.0

# Kaggle API
kaggle>=1.5.16

# Quantum Computing
pennylane>=0.33.0
pennylane-lightning>=0.33.0

